fix: Avoid writing event handler wrapper code to disk
